+++ src/main.ts
@@ -2,3 +2,4 @@ export * from "./datetime";
 export * from "./duration";
 export * from "./interval";
 export * from "./number";
+export * from "./string";